+++ computer-confugration-website/src/app/complete-build/page.jsx
@@ -1,0 +1,157 @@
+import React from 'react'
+import '@/styles/completebuild.css'
+import picture from '@/img/pic1.jpg'
+import Image from 'next/image'
+
+const completebuild = () => {
+  return (
+    <>
+        <section className="page-title">
+            <h1>Complete PC builds</h1>
+        </section>
+
+        <section className="filter-section">
+            <button className="filter-btn">Filters</button>
+            <input type="text" className="search-input" placeholder="Hinted search text"></input>
+        </section>
+
+            <section className="container">
+            <section className="builds-gallery">
+            <div className="build-item">
+                <Image src={picture} alt='PC build'></Image>
+                <div className="build-info">
+                    <h2>Exceptional Black Build with Core i7-12700F</h2>
+                    <p>Intel Core i7-12700F | GeForce RTX 3060 12GB</p>
+                    <p>16GB DDR4 3200 | 500GB SSD</p>
+                    <p>LGA170 Semi-Modular 600W</p>
+                    <button className="btn">Price: $1500</button>
+                </div>
+            </div>
+            <div className="build-item">
+                <Image src={picture} alt='PC build'></Image>
+                <div className="build-info">
+                    <h2>Exceptional Black Build with Core i7-12700F</h2>
+                    <p>Intel Core i7-12700F | GeForce RTX 3060 12GB</p>
+                    <p>16GB DDR4 3200 | 500GB SSD</p>
+                    <p>LGA170 Semi-Modular 600W</p>
+                    <button className="btn">Price: $1500</button>
+                </div>
+            </div>
+            <div className="build-item">
+                <Image src={picture} alt='PC build'></Image>
+                <div className="build-info">
+                    <h2>Exceptional Black Build with Core i7-12700F</h2>
+                    <p>Intel Core i7-12700F | GeForce RTX 3060 12GB</p>
+                    <p>16GB DDR4 3200 | 500GB SSD</p>
+                    <p>LGA170 Semi-Modular 600W</p>
+                    <button className="btn">Price: $1500</button>
+                </div>
+            </div>
+            <div className="build-item">
+                <Image src={picture} alt='PC build'></Image>
+                <div className="build-info">
+                    <h2>Exceptional Black Build with Core i7-12700F</h2>
+                    <p>Intel Core i7-12700F | GeForce RTX 3060 12GB</p>
+                    <p>16GB DDR4 3200 | 500GB SSD</p>
+                    <p>LGA170 Semi-Modular 600W</p>
+                    <button className="btn">Price: $1500</button>
+                </div>
+            </div>
+            <div className="build-item">
+                <Image src={picture} alt='PC build'></Image>
+                <div className="build-info">
+                    <h2>Exceptional Black Build with Core i7-12700F</h2>
+                    <p>Intel Core i7-12700F | GeForce RTX 3060 12GB</p>
+                    <p>16GB DDR4 3200 | 500GB SSD</p>
+                    <p>LGA170 Semi-Modular 600W</p>
+                    <button className="btn">Price: $1500</button>
+                </div>
+            </div>
+            <div className="build-item">
+                <Image src={picture} alt='PC build'></Image>
+                <div className="build-info">
+                    <h2>Exceptional Black Build with Core i7-12700F</h2>
+                    <p>Intel Core i7-12700F | GeForce RTX 3060 12GB</p>
+                    <p>16GB DDR4 3200 | 500GB SSD</p>
+                    <p>LGA170 Semi-Modular 600W</p>
+                    <button className="btn">Price: $1500</button>
+                </div>
+            </div>
+            <div className="build-item">
+                <Image src={picture} alt='PC build'></Image>
+                <div className="build-info">
+                    <h2>Exceptional Black Build with Core i7-12700F</h2>
+                    <p>Intel Core i7-12700F | GeForce RTX 3060 12GB</p>
+                    <p>16GB DDR4 3200 | 500GB SSD</p>
+                    <p>LGA170 Semi-Modular 600W</p>
+                    <button className="btn">Price: $1500</button>
+                </div>
+            </div>
+            <div className="build-item">
+                <Image src={picture} alt='PC build'></Image>
+                <div className="build-info">
+                    <h2>Exceptional Black Build with Core i7-12700F</h2>
+                    <p>Intel Core i7-12700F | GeForce RTX 3060 12GB</p>
+                    <p>16GB DDR4 3200 | 500GB SSD</p>
+                    <p>LGA170 Semi-Modular 600W</p>
+                    <button className="btn">Price: $1500</button>
+                </div>
+            </div>
+            <div className="build-item">
+                <Image src={picture} alt='PC build'></Image>
+                <div className="build-info">
+                    <h2>Exceptional Black Build with Core i7-12700F</h2>
+                    <p>Intel Core i7-12700F | GeForce RTX 3060 12GB</p>
+                    <p>16GB DDR4 3200 | 500GB SSD</p>
+                    <p>LGA170 Semi-Modular 600W</p>
+                    <button className="btn">Price: $1500</button>
+                </div>
+            </div>
+            <div className="build-item">
+                <Image src={picture} alt='PC build'></Image>
+                <div className="build-info">
+                    <h2>Exceptional Black Build with Core i7-12700F</h2>
+                    <p>Intel Core i7-12700F | GeForce RTX 3060 12GB</p>
+                    <p>16GB DDR4 3200 | 500GB SSD</p>
+                    <p>LGA170 Semi-Modular 600W</p>
+                    <button className="btn">Price: $1500</button>
+                </div>
+            </div>
+            <div className="build-item">
+                <Image src={picture} alt='PC build'></Image>
+                <div className="build-info">
+                    <h2>Exceptional Black Build with Core i7-12700F</h2>
+                    <p>Intel Core i7-12700F | GeForce RTX 3060 12GB</p>
+                    <p>16GB DDR4 3200 | 500GB SSD</p>
+                    <p>LGA170 Semi-Modular 600W</p>
+                    <button className="btn">Price: $1500</button>
+                </div>
+            </div>
+            <div className="build-item">
+                <Image src={picture} alt='PC build'></Image>
+                <div className="build-info">
+                    <h2>Exceptional Black Build with Core i7-12700F</h2>
+                    <p>Intel Core i7-12700F | GeForce RTX 3060 12GB</p>
+                    <p>16GB DDR4 3200 | 500GB SSD</p>
+                    <p>LGA170 Semi-Modular 600W</p>
+                    <button className="btn">Price: $1500</button>
+                </div>
+            </div>
+            </section>
+            </section>
+        <div className="pagination">
+            <button className="pagination-btn">Previous</button>
+            <div className="page-numbers">
+                <button className="page-number active">1</button>
+                <button className="page-number">2</button>
+                <button className="page-number">3</button>
+                <span>...</span>
+                <button className="page-number">68</button>
+            </div>
+            <button className="pagination-btn">Next</button>
+        </div> 
+    </>
+  )
+}
+
+export default completebuild
