--- computer-confugration-website/src/app/complete_build/build_data.js
@@ -1,239 +1,0 @@
-import picture from '@/img/pic1.jpg';
-
-const products = [
-    {
-      id: 1,
-      picture: picture,
-      title: "Exceptional Black Build with Core i7-12700F",
-      specs: [
-        "Intel Core i7-12700F | GeForce RTX 3060 12GB",
-        "16GB DDR4 3200 | 500GB SSD",
-        "LGA170 Semi-Modular 600W",
-      ],
-      price: 1500,
-    },
-    {
-        id: 2,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 3,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 4,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 5,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 6,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 7,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 8,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 9,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 10,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 11,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 12,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 13,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 14,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 15,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 16,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 17,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-    {
-        id: 18,
-        picture: "/img/pic1.jpg",
-        title: "Exceptional Black Build with Core i7-12700F",
-        specs: [
-          "Intel Core i7-12700F",
-          "GeForce RTX 3060 12GB",
-          "16GB DDR4 3200",
-          "500GB SSD",
-          "LGA170 Semi-Modular 600W",
-        ],
-        price: 1500,
-    },
-];
-  
-export default products;
-
