--- computer-confugration-website/src/styles/builddetail.css
@@ -1,74 +1,0 @@
-.main_content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    background-color: #F7FFF7;
-}
-
-.page_title {
-    height: auto;
-    width: 100%;
-    background-color: #14b8a6;
-    text-align: center;
-    padding: 2rem 0;
-}
-
-.page_title h4 {
-    font-size: 1rem;
-}
-
-.build_name {
-    color: #1f2937;
-    font-size: 2rem;
-}
-
-.user {
-    color: #F7FFF7;
-}
-
-.page_content {
-    width: 1100px;
-    min-height: 100vh;
-    color: #1f2937;
-    padding: 2rem 0;
-    display: block;
-}
-
-.sidebar {
-    width: 33%;
-    background-color: blue;
-    float: left;
-}
-
-.mainbar {
-    width: 66%;
-    background-color: green;
-    float: right;
-    padding-left: 1rem;
-    padding-top: 1rem;
-}
-
-.description_box {
-    margin-bottom: 2rem;
-}
-
-.input_comment {
-    height: 400px;
-    width: 70%;
-    margin-top: 1rem;
-    margin-bottom: 1rem;
-}
-
-.cmbox {
-    width: 100%;
-    height: 300px;
-}
-
-.savebtn {
-    background-color: #14b8a6;
-    border: 1px #1f2937;
-    margin: 1rem 0;
-    padding: 10px 20px;
-    border-radius: 10%;
-}
