--- computer-confugration-website/src/styles/test.css
@@ -1,37 +1,0 @@
-/* styles/globals.css */
-.build-item {
-    border: 1px solid #ddd;
-    border-radius: 8px;
-    padding: 16px;
-    text-align: center;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-  }
-  
-  .build-info {
-    margin-top: 12px;
-  }
-  
-  .build-info h2 {
-    font-size: 18px;
-    margin-bottom: 8px;
-  }
-  
-  .build-info p {
-    margin: 4px 0;
-    color: #555;
-  }
-  
-  .btn {
-    margin-top: 12px;
-    padding: 8px 16px;
-    background-color: #0070f3;
-    color: #fff;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-  }
-  
-  .btn:hover {
-    background-color: #005bb5;
-  }
-
